pyproject.toml
[project]
name = "mlops-streaming-project"
version = "0.1.0"
description = "Streaming ML pipeline (MLOps MVP)"
requires-python = ">=3.10"

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "tqdm"
]